requirements.txt
numpy == 1.21.4
opencv-python == 4.5.4.58
Pillow == 8.4.0
tensorflow == 2.7.0
scikit-image == 0.18.3
imutils == 0.5.4
